+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToStream/loadable.ts
@@ -1,7 +1,7 @@
 import { resolve } from 'path';
 import { ChunkExtractor } from '@loadable/server';
 import { LOADABLE_STATS_FILE } from '@modern-js/utils';
-import { RuntimeContext } from '../../core';
+import { RuntimeContext } from '../types';
 
 export type LoadableChunk = {
   filename?: string;
@@ -11,18 +11,13 @@ export function getLoadableChunks(
   jsx: React.ReactElement,
 ) {
   const { ssrContext } = context;
-  if (!ssrContext) {
-    // TODO: handle Error;
-    throw new Error('has not ssrContext');
-  }
-
-  const loadableManifest = resolve(ssrContext.distDir, LOADABLE_STATS_FILE);
+  const loadableManifest = resolve(ssrContext!.distDir, LOADABLE_STATS_FILE);
   if (!loadableManifest) {
     return [];
   }
   const extractor = new ChunkExtractor({
     statsFile: loadableManifest,
-    entrypoints: [ssrContext.entryName],
+    entrypoints: [ssrContext!.entryName],
   });
   extractor.collectChunks(jsx);
   const chunks = extractor.getChunkAssets(extractor.chunks) || [];
