+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToStream/styledComponent.ts
@@ -1,11 +1,12 @@
 import { ServerStyleSheet } from 'styled-components';
-import { RuntimeContext } from '../../core';
+import { RuntimeContext } from '../types';
 
 export function getStyledComponentCss(
   _context: RuntimeContext,
   jsx: React.ReactElement,
 ) {
   const sheet = new ServerStyleSheet();
+
   sheet.collectStyles(jsx);
   return sheet.getStyleTags();
 }
