+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToStream/type.ts
