+++ packages/runtime/plugin-runtime/src/ssr/serverRender/types.ts
@@ -1,0 +1,9 @@
+import type { RuntimeContext } from '../../core';
+import { RenderLevel } from './renderToString/type';
+
+export type { SSRServerContext } from './renderToString/type';
+export type ModernSSRReactComponent = React.ComponentType<any> & {
+  init: (context: RuntimeContext) => Promise<void>;
+  prefetch: (context: RuntimeContext) => Promise<Record<string, any>>;
+};
+export { RuntimeContext, RenderLevel };
